improviments on ocl query generator

+++ faultTypeRuleExecutor-workspace/faultTypeRuleExecutor/src/command/InjectionCommandFactory.java
@@ -1,11 +1,19 @@
 package command;
 
-import command.impl.DeleteBlockContainerRelation;
+import command.impl.MiaAction;
+import command.impl.MiebAction;
 import command.impl.DeleteNodeCommand;
 
 public final class InjectionCommandFactory {
 
 	public static InjectionCommand createsAction(String actionName) {
-		return new DeleteNodeCommand();
+		if (actionName.toLowerCase().equals("delete")) {
+			return new DeleteNodeCommand();
+		} else if (actionName.toLowerCase().equals("mia")) {
+			return new MiaAction();
+		} else if (actionName.toLowerCase().equals("mieb")) {
+			return new MiebAction();
+		}
+		return null;
 	}
 }
